move test out of utils
